refactor: Login 페이지를 SingleSectionLayout, LoginTemplate으로 분할
