--- src/routes/path.ts
@@ -1,4 +1,0 @@
-export const PATH = {
-  home: '/',
-  login: '/login',
-} as const;
